+++ lib/asset/baseAsset.js
@@ -62,8 +62,7 @@ class BaseAsset {
 
     /**
      * 20-byte hex ID of the document
-     * @type string
-     * @readonly
+     * @returns {String} asset_id - The Asset ID
      */
     get asset_id() {
         return this._asset_id;
@@ -87,8 +86,8 @@ class BaseAsset {
 
     /**
      * Sets an image asset to be the article's thumbnail. (An image asset is
-     * returned e.g. from {@link utildownload_img|util.download_img} or
-     * {@link utildownload_image|util.download_image}.)
+     * returned e.g. from {@link #utildownloadimagefromimgtag|util.downloadImg}
+     * or {@link #utildownloadimage|util.downloadImage}.)
      * @param {ImageAsset} value - Image
      */
     set_thumbnail(value) {
