+++ lib/asset/imageAsset.js
@@ -11,8 +11,8 @@ class ImageAsset extends BaseAsset {
     }
 
     /**
-     * @param {string} content_type
-     * @param {(Buffer|Promise<Buffer>)} image_data
+     * @param {string} contentType
+     * @param {(Buffer|Promise<Buffer>)} imageData
      */
     set_image_data(contentType, imageData) {
         this._content_type = contentType;
