+++ lib/util.js
@@ -5,7 +5,6 @@
 
 const cheerio = require('cheerio');
 const crypto = require('crypto');
-const Feedparser = require('feedparser');
 const parseDataUrl = require('parse-data-url');
 const request = require('request');
 const url = require('url');
@@ -15,6 +14,8 @@ const { logger } = require('./logging');
 const somaDOM = require('./soma_dom');
 const ImageAsset = require('./asset/imageAsset');
 const VideoAsset = require('./asset/videoAsset');
+const rss = require('./util/rss');
+const { delayExecution, get_ingestion_limits } = require('./util/common');
 
 const USER_AGENT = 'libingester';
 
@@ -57,12 +58,6 @@ function promisifiedRequest(options) {
 
             return resolve(res);
         });
-    });
-}
-
-function delayExecution(timeout) {
-    return new Promise(resolve => {
-        setTimeout(resolve, timeout);
     });
 }
 
@@ -141,9 +136,9 @@ function fetchHtml(uri, encoding) {
 
 /**
  * @param {Cheerio} doc - DOM tree of an HTML document
- * @param {string} base_uri - Fallback base URI, should be the document's URI
+ * @param {string} baseUri - Fallback base URI, should be the document's URI
  * @returns {string} - Base URI specified by the document if there is one,
- *   otherwise `base_uri`
+ *   otherwise `baseUri`
  * @memberof util
  */
 function getDocBaseUri(doc, baseUri) {
@@ -157,9 +152,9 @@ function getDocBaseUri(doc, baseUri) {
 }
 
 /**
- * @param {Cheerio} video_tag - DOM element to replace with a placeholder for
+ * @param {Cheerio} videoTag - DOM element to replace with a placeholder for
  *   the video to be downloaded later
- * @param {string} video_uri - URI from which to download the video
+ * @param {string} videoUri - URI from which to download the video
  * @returns {VideoAsset}
  * @memberof util
  */
@@ -244,9 +239,9 @@ function downloadImage(uri) {
 }
 
 /**
- * Downloads a thumbnail image for the Youtube video specified by `$embed_url`.
+ * Downloads a thumbnail image for the Youtube video specified by `$embedUrl`.
  *
- * @param {string} embed_url - Youtube embed video URL
+ * @param {string} embedUrl - Youtube embed video URL
  * @returns {ImageAsset}
  * @memberof util
  */
@@ -273,10 +268,10 @@ function downloadYoutubeThumbnail(embedUrl) {
  * This is a convenience function which saves you from having to extract the
  * image URI yourself and pass it to {@link downloadImage}.
  *
- * Pass `base_uri` to handle relative links in the `src` attribute.
+ * Pass `baseUri` to handle relative links in the `src` attribute.
  *
  * @param {Cheerio} $img - DOM tree of an `<img>` element
- * @param {string} base_uri - Base URI of the document
+ * @param {string} baseUri - Base URI of the document
  * @returns {ImageAsset}
  * @memberof util
  */
@@ -331,227 +326,20 @@ function downloadImageFromImgTag($img, baseUri = '') {
     return asset;
 }
 
-/**
- * Get parsed settings to limit the ingested items.
- *
- * The `maxItems` and `maxDaysOld` parameters can be overridden by the
- * `LIBINGESTER_MAX_ITEMS` and `LIBINGESTER_MAX_DAYS_OLD` environment
- * variables, respectively.
- *
- * @param {number} [desiredMaxItems=Infinity] - Maximum number of entries to ingest
- * @param {number} [desiredMaxDaysOld=1] - Maximum age of entries to ingest, in days
- * @returns {Object} - Object with parsed settings `max_items` and `oldest_date`
- * @memberof util
- */
-function getIngestionLimits(desiredMaxItems = Infinity, desiredMaxDaysOld = 1) {
-    const actualMaxDaysOld = process.env.LIBINGESTER_MAX_DAYS_OLD || desiredMaxDaysOld;
-    const oldestDate = new Date();
-    oldestDate.setDate(oldestDate.getDate() - parseInt(actualMaxDaysOld, 10));
-
-    let actualMaxItems = process.env.LIBINGESTER_MAX_ITEMS || desiredMaxItems;
-    if (actualMaxItems !== Infinity) {
-        actualMaxItems = parseInt(actualMaxItems, 10);
-    }
-
-    return {
-        max_items: actualMaxItems,
-        oldest_date: oldestDate,
-    };
-}
-
-function _fetchRssPage(feed, page, items, allLinks, maxItems, oldestDate) {
-    let isPaginated = false;
-    let feedUrl;
-
-    let feedObj = feed;
-    if (Array.isArray(feed)) {
-        feedObj = feed[0];
-    }
-
-    if (typeof feedObj === 'function') {
-        isPaginated = true;
-        feedUrl = feedObj(page);
-    } else {
-        feedUrl = feedObj;
-    }
-
-    if (!feedUrl) {
-        return items;
-    }
-
-    logger.debug(`Fetching RSS ${feedUrl}`);
-
-    return _fetchRssJsonWithRetry(feedUrl).then(feedJson => {
-        const recentEnoughItems = feedJson.items
-              .filter(item => item.pubdate >= oldestDate);
-
-        const limitedItems = [];
-        recentEnoughItems.slice(0, maxItems - items.length).forEach(item => {
-            if (!allLinks.has(item.link)) {
-                limitedItems.push(item);
-                allLinks.add(item.link);
-            }
-        });
-
-        const newItems = items.concat(limitedItems);
-
-        // If we've run into articles which are too old, or we've hit the max
-        // number of items, cease crawling
-        const doneCrawling = feedJson.items.length === 0 ||
-                             limitedItems.length < feedJson.items.length;
-
-        if (typeof feedObj !== 'function' && _isWordpressFeed(feedJson)) {
-            logger.debug('Wordpress feed found, adding pagination');
-            isPaginated = true;
-            const newFeed = _doCreateWordpress(feedObj);
-            if (Array.isArray(feed)) {
-                feed.shift();
-                // eslint-disable-next-line no-param-reassign
-                feed = [newFeed].concat(feed);
-            } else {
-                // eslint-disable-next-line no-param-reassign
-                feed = newFeed;
-            }
-        }
-
-        // eslint-disable-next-line no-extra-parens
-        const _continue = Array.isArray(feed) || (!doneCrawling && isPaginated);
-
-        if (!_continue) {
-            return newItems;
-        }
-
-        let newPage;
-        if (doneCrawling && isPaginated && Array.isArray(feed)) {
-            feed.shift();
-            newPage = 1;
-        } else if (isPaginated) {
-            newPage = page + 1;
-        } else if (Array.isArray(feed)) {
-            feed.shift();
-            newPage = 1;
-        }
-        return _fetchRssPage(feed, newPage, newItems, allLinks, maxItems, oldestDate);
-    });
-}
-
-/**
- * A utility for fetching entries from an RSS feed for further processing.
- *
- * The `max_items` and `max_days_old` parameters can be overridden by the
- * `LIBINGESTER_MAX_ITEMS` and `LIBINGESTER_MAX_DAYS_OLD` environment
- * variables, respectively.
- *
- * @param {string|Array|Function} feed - URL of RSS feed to read, or an array of
- *   URLs of RSS feeds to read, or a paginator function that returns a URL given
- *   a page number
- * @param {number} [max_items=Infinity] - Maximum number of entries to return
- * @param {number} [max_days_old=1] - Maximum age of entries returned, in days
- * @returns {Promise<Object[]>} - Array of RSS feed entries
- * @memberof util
- */
-function fetchRssEntries(feed, max_items = Infinity, max_days_old = 1) {
-    // Allow environment variables to override this
-    const limits = getIngestionLimits(max_items, max_days_old);
-
-    return _fetchRssPage(feed, 1, [], new Set(), limits.max_items,
-                         limits.oldest_date);
-}
-
-function _fetchRssJson(feedUrl) {
-    return new Promise((resolve, reject) => {
-        const req = request.get(feedUrl);
-        const parser = new Feedparser();
-
-        req.on('error', reject);
-        parser.on('error', reject);
-
-        const feedData = { items: [] };
-
-        parser.on('meta', meta => {
-            feedData.meta = meta;
-        });
-
-        parser.on('data', article => {
-            feedData.items.push(article);
-        });
-
-        parser.on('end', () => resolve(feedData));
-
-        req.pipe(parser);
-    });
-}
-
-function _isWordpressFeed(feedJson) {
-    const hasGenerator = feedJson.meta.generator &&
-          feedJson.meta.generator.startsWith('https://wordpress.org/');
-    const hasXmlns = feedJson.meta['rss:site'] &&
-          feedJson.meta['rss:site']['@'] &&
-          feedJson.meta['rss:site']['@'].xmlns &&
-          feedJson.meta['rss:site']['@'].xmlns.startsWith('com-wordpress');
-
-    return hasGenerator || hasXmlns;
-}
-
-function _fetchRssJsonWithRetry(feedUrl, attempt = 1) {
-    const maxRetries = parseInt(config.get_setting('max-retries'), 10);
-    const retryBackoffDelay = parseInt(config.get_setting('retry-backoff-delay'), 10);
-
-    return _fetchRssJson(feedUrl).catch(err => {
-        if (attempt > maxRetries) {
-            logger.warn(`Max retries (${maxRetries}) exceeded!`);
-            throw err;
-        }
-
-        const shouldRetry = err.message.match('Unexpected end') ||
-                            err.message.match('Not a feed');
-        if (shouldRetry) {
-            const timeout = Math.pow(2, attempt) * retryBackoffDelay;
-            logger.debug(`Delaying execution of ${feedUrl} by ${timeout}`);
-
-            return delayExecution(timeout).then(() => {
-                return _fetchRssJsonWithRetry(feedUrl, attempt + 1);
-            });
-        }
-
-        // We reach here if unexpected error
-        throw err;
-    });
-}
-
-function _doCreateWordpress(feed) {
-    return function (pageNum) {
-        return `${feed}?paged=${pageNum}`;
-    };
-}
-
-/**
- * A utility to create a pagination function for Wordpress-generated feeds,
- * since they are so common. Make sure to verify that your feed is generated by
- * Wordpress before using this!
- * @param {string|Array} feed - The URL of the RSS feed (first page), or an
- *   array of URLs of RSS feeds
- * @returns {function} - A paginator function that can be passed to
- *   {@link util.fetchRssEntries|util.fetchRssEntries}
- * @memberof util
- */
-function createWordpressPaginator(feed) {
-    if (Array.isArray(feed)) {
-        return feed.map(uri => _doCreateWordpress(uri));
-    }
-    return _doCreateWordpress(feed);
-}
 
 module.exports = {
-    create_wordpress_paginator: createWordpressPaginator,
     download_image: downloadImage,
     download_img: downloadImageFromImgTag,
     download_youtube_thumbnail: downloadYoutubeThumbnail,
     encode_uri: encodeUri,
     fetch_html: fetchHtml,
-    fetch_rss_entries: fetchRssEntries,
     get_doc_base_uri: getDocBaseUri,
     get_embedded_video_asset: getEmbeddedVideoAsset,
-    get_ingestion_limits: getIngestionLimits,
+    get_ingestion_limits,
     set_user_agent: setUserAgent,
+    rss,
+
+    // For backwards compatibility:
+    create_wordpress_paginator: rss.create_wordpress_paginator,
+    fetch_rss_entries: rss.fetch_rss_entries,
 };
